backend/requirements.txt
fastapi
uvicorn[standard]
google-genai>=1.0.0
google-adk>=0.1.0
google-cloud-firestore>=2.16.0
google-cloud-speech>=2.26.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
Pillow>=10.2.0
httpx>=0.27.0
python-multipart>=0.0.9
